MAke dashboard policies work
